app/requirements.txt
# Streamlit Dashboard Requirements
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
